PDFBOX-3017: move two classes to cert package
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1846747 13f79535-47bb-0310-9956-ffa450edef68
